pyproject.toml
[project]
name = "fastapi-postgres-demo"
version = "0.1.0"
description = "A simple FastAPI application that connects to a PostgreSQL database"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.11",
    "uvicorn==0.34.0",
    "sqlalchemy==2.0.38",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 